LayoutTests/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A5_T2-expected.txt

CONSOLE MESSAGE: line 82: SyntaxError: Label 'IN_DO_FUNC' is not defined
S12.8_A5_T2

PASS Expected parsing failure

TEST COMPLETE
